feat(MAVLink): support 32-bit system IDs
Carry full-width system IDs through vehicle discovery, message targeting, settings, and signing without changing MAVLink payload layouts. Patch the pinned C generator and libevents to match the experimental ArduPilot and pymavlink extensions while preserving legacy framing for ordinary IDs.

Cover wide vehicle and GCS IDs, parameter and FTP exchanges, independent sequence tracking, and signed forwarding with regression tests.

+++ src/AnalyzeView/MAVLinkInspector/MAVLinkInspectorController.h
@@ -50,7 +50,7 @@ public:
     explicit MAVLinkInspectorController(QObject *parent = nullptr);
     ~MAVLinkInspectorController();
 
-    Q_INVOKABLE void setActiveSystem(int systemId);
+    Q_INVOKABLE void setActiveSystem(quint32 systemId);
     Q_INVOKABLE void setMessageInterval(int32_t rate) const;
 
     QmlObjectListModel *systems() const { return _systems; }
@@ -76,8 +76,8 @@ private slots:
     void _vehicleRemoved(const Vehicle *vehicle);
 
 private:
-    QGCMAVLinkSystem *_findVehicle(uint8_t id);
-    uint8_t _selectedSystemID() const;
+    QGCMAVLinkSystem *_findVehicle(quint32 id);
+    quint32 _selectedSystemID() const;
     uint8_t _selectedComponentID() const;
 
     QStringList _timeScales;
