+++ src/AnalyzeView/MAVLinkInspector/MAVLinkMessage.h
@@ -34,7 +34,7 @@ public:
     static QString extractInstanceValue(const mavlink_message_t &message);
 
     quint32 id() const { return _message.msgid;  }
-    quint8 sysId() const { return _message.sysid; }
+    quint32 sysId() const { return _message.sysid; }
     quint8 compId() const { return _message.compid; }
     QString name() const { return _name;  }
     QString instanceValue() const { return _instanceValue; }
